% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_outcome_sql.R
\name{create_outcome_sql}
\alias{create_outcome_sql}
\title{Create Sql Code for outcome (adverse drug effect) variables}
\usage{
create_outcome_sql(
  concept_id,
  concept_name,
  filepath = "~/HowOften2023/inst/sql/sql_server/outcome/"
)
}
\arguments{
\item{concept_id}{integer, SNOMED CT}

\item{concept_name}{string, name of the side effect}

\item{filepath}{string, filepath of where you want the sql files to be}
}
\value{
SQL file that can be run to generate cohorts
}
\description{
Create Sql Code for outcome (adverse drug effect) variables
}
\examples{
create_outcome_sql(8941340, "Name of Outcome")
}
